fix(middleware): cast duration_ms to numeric in site_access_events insert
PostgreSQL cannot infer the type of a bare numeric parameter inside
jsonb_build_object(); adding ::numeric resolves the ambiguity and
fixes the 503 that was hitting every route.

+++ backend/auth/middleware/trackSiteAccess.js
@@ -57,7 +57,7 @@ function trackSiteAccess(opts = {}) {
               now(), $1, $2, $3,
               $4, $5,
               $6::inet, digest($6::text, 'sha256'), $7, $8,
-              jsonb_build_object('duration_ms', $9)
+              jsonb_build_object('duration_ms', $9::numeric)
             )
             `,
             [pathname, method, statusCode, userId, sessionId, ip, userAgent, referer, durationMs]
